Move check-font-loaded into utils
