+++ src/utils/check-font-loaded.js
